[SPARK-15115][SQL] Reorganize whole stage codegen benchmark suites
